Fix #29: Visitor pattern
Basically refactoring the code a bit to ease the use of methods
